+++ python/tests/collections/volume_test.py
@@ -47,6 +47,26 @@ test_cases_volume_xml = (
   </frontmatter>
 </volume>
 """,
+    """<volume id="demo" type="proceedings" ingest-date="2022-05-15">
+  <meta>
+    <booktitle>Proceedings of the 60th Annual Meeting of the Association for Computational Linguistics: System Demonstrations</booktitle>
+    <editor><first>Valerio</first><last>Basile</last></editor>
+    <editor><first>Zornitsa</first><last>Kozareva</last></editor>
+    <editor><first>Sanja</first><last>Stajner</last></editor>
+    <publisher>Association for Computational Linguistics</publisher>
+    <address>Dublin, Ireland</address>
+    <doi>10.18653/v1/2022.acl-demo</doi>
+    <month>May</month>
+    <year>2022</year>
+    <url hash="d92e3f4d">2022.acl-demo</url>
+    <venue>acl</venue>
+  </meta>
+  <frontmatter>
+    <url hash="ad64a7d9">2022.acl-demo.0</url>
+    <bibkey>acl-2022-association-linguistics-system</bibkey>
+  </frontmatter>
+</volume>
+""",
     """<volume id="1" type="journal">
   <meta>
     <booktitle>Computational Linguistics, Volume 15, Number 1, March 1989</booktitle>
@@ -138,7 +158,7 @@ def test_volume_all_attribs():
     assert volume.get_ingest_date() == date(2023, 1, 12)
 
 
-def test_volume_attributes_2022acl(anthology):
+def test_volume_attributes_2022acl_long(anthology):
     volume = anthology.get_volume("2022.acl-long")
     assert isinstance(volume, Volume)
     assert volume.id == "long"
@@ -146,11 +166,31 @@ def test_volume_attributes_2022acl(anthology):
     assert volume.get_ingest_date() == date(2022, 5, 15)
     assert volume.address == "Dublin, Ireland"
     assert volume.publisher == "Association for Computational Linguistics"
+    assert volume.doi is None
     assert volume.month == "May"
     assert volume.year == "2022"
     assert volume.pdf.name == "2022.acl-long"
     assert volume.pdf.checksum == "b8317652"
     assert volume.venue_ids == ["acl"]
+    assert volume.venue_acronym == "ACL"
+    assert not volume.is_workshop
+
+
+def test_volume_attributes_2022acl_demo(anthology):
+    volume = anthology.get_volume("2022.acl-demo")
+    assert isinstance(volume, Volume)
+    assert volume.id == "demo"
+    assert volume.ingest_date == "2022-05-15"
+    assert volume.get_ingest_date() == date(2022, 5, 15)
+    assert volume.address == "Dublin, Ireland"
+    assert volume.publisher == "Association for Computational Linguistics"
+    assert volume.doi == "10.18653/v1/2022.acl-demo"
+    assert volume.month == "May"
+    assert volume.year == "2022"
+    assert volume.pdf.name == "2022.acl-demo"
+    assert volume.pdf.checksum == "d92e3f4d"
+    assert volume.venue_ids == ["acl"]
+    assert volume.venue_acronym == "ACL"
     assert not volume.is_workshop
 
 
@@ -159,6 +199,7 @@ def test_volume_attributes_j89(anthology):
     assert isinstance(volume, Volume)
     assert volume.id == "1"
     assert volume.venue_ids == ["cl"]
+    assert volume.venue_acronym == "CL"
     assert volume.year == "1989"
     assert not volume.is_workshop
     assert volume.type == VolumeType.JOURNAL
@@ -173,6 +214,8 @@ def test_volume_attributes_naloma(anthology):
     assert volume.id == "1"
     assert volume.year == "2022"
     assert volume.is_workshop
+    assert volume.venue_ids == ["nlma", "ws"]
+    assert volume.venue_acronym == "NALOMA"
 
 
 def test_volume_venues_j89(anthology):
@@ -185,10 +228,11 @@ def test_volume_venues_j89(anthology):
 
 def test_volume_venues_naloma(anthology):
     volume = anthology.get_volume("2022.naloma-1")
-    assert volume.venue_ids == ["nlma"]
+    assert volume.venue_ids == ["nlma", "ws"]
     venues = volume.venues()
-    assert len(venues) == 1
+    assert len(venues) == 2
     assert venues[0].id == "nlma"
+    assert venues[1].id == "ws"
 
 
 def test_volume_with_nonexistent_venue(anthology):
@@ -206,9 +250,35 @@ def test_volume_with_nonexistent_venue(anthology):
         _ = volume.venues()
 
 
+def test_volume_with_multiple_venues(anthology):
+    volume_title = MarkupText.from_string(
+        "Joint conference of ACL and LREC (hypothetical)"
+    )
+    parent = Collection("acl.2092", CollectionIndexStub(anthology), Path("."))
+    volume = Volume(
+        "1",
+        parent,
+        type=VolumeType.PROCEEDINGS,
+        booktitle=volume_title,
+        venue_ids=["acl", "lrec"],
+        year="2092",
+    )
+    assert volume.full_id == "acl.2092-1"
+    assert volume.title == volume_title
+    assert volume.venue_ids == ["acl", "lrec"]
+    assert volume.venue_acronym == "ACL-LREC"
+
+
 def test_volume_get_events(anthology):
     volume = anthology.get_volume("2022.acl-demo")
     assert volume.get_events() == [anthology.events["acl-2022"]]
+
+
+def test_volume_get_sigs(anthology):
+    volume = anthology.get_volume("2022.acl-demo")
+    assert volume.get_sigs() == [anthology.sigs["sigdat"]]
+    volume = anthology.get_volume("2022.acl-long")
+    assert volume.get_sigs() == []
 
 
 @pytest.mark.parametrize("xml", test_cases_volume_xml)