+++ python/CHANGELOG.md
@@ -2,6 +2,55 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support for Python 3.13.
+
+## [0.5.1] — 2025-01-02
+
+This release changes the PyPI package name from acl-anthology-py to acl-anthology.
+
+### Added
+
+- VenueIndex can now set `no_item_ids=True` to skip reverse-indexing volumes. This avoids parsing all XML files if all you want to access is basic venue information, but means that `Venue.item_ids` will be empty. _You probably don't want to use this unless you know that you are not going to need this information._
+
+### Changed
+
+- LaTeX encoding now uses [pylatexenc](https://pylatexenc.readthedocs.io/) instead of latexcodec, and wraps all macros in braces. This should address problems with BibTeX handling, see [#4280](https://github.com/acl-org/acl-anthology/issues/4280).
+
+## [0.5.0] — 2024-12-25
+
+This release is intended to be feature-complete with regard to generating the entire ACL Anthology website.
+
+### Added
+
+- Papers can now generate citation reference strings via `to_citation()`.
+  - Calling `to_citation()` without any arguments will produce ACL-formatted reference entries.
+  - Alternatively, `to_citation()` can be called with the path to a CSL style file, in which case it will use citeproc-py to generate an entry formatted to the specifications in that style file.
+- Papers can now generate brief markdown reference strings via `to_markdown_citation()`.
+- PersonIndex now has function `find_coauthors_counter()` to find not just the identities of co-authors, but also get a count how many items they have co-authored together with someone.
+- SIGIndex now reverse-indexes co-located volumes, so it is now possible to get SIGs associated with volumes, e.g. via `Volume.get_sigs()`.
+- VenueIndex now reverse-indexes associated volumes, so it is now possible to get volumes associated with venues, e.g. via `Venue.volumes()`.
+- Papers now have attribute `thumbnail`.
+- Papers now have attribute `language_name`, which uses the [langcodes](https://langcodes-hickford.readthedocs.io/en/) library to map language tags in the XML to proper language names.
+- Papers now have attributes `issue` and `journal` for edge cases where these are set on the paper level. `Paper.get_issue()` and `Paper.get_journal_title()` can be used to access them without having to know where they are defined.
+- Volumes now have attributes `has_abstracts`, `venue_acronym`, and `web_url`.
+- Names now have function `as_full()`, returning the full name in the appropriate format based on whether it is given in Han or Latin script.
+- MarkupText now has function `as_xml()` to return a string of the internal XML representation.
+
+### Changed
+
+- `Venue.item_ids` and `Person.item_ids` are now lists instead of sets. This is because we need to preserve the order in which items were added when loading the XML, as this is meaningful (e.g. reflects the order in which items should appear on the Anthology website).
+- `Paper.attachments` is now a list of tuples, instead of a dict. This is because attachment types are not always unique (e.g., there can be two "software" attachments).
+- Bugfix: Events now use the correct URL template.
+- Bugfix: Events that are both implicitly _and_ explicitly created now merge their information, instead of overwriting each other.
+- Bugfix: Converting a `<texmath>` expression to Unicode no longer serializes the tail of the XML tag, but only the TeX math expression itself.
+- Bugfix: Heuristic scoring of name variants will no longer overwrite canonical names that are explicitly defined in `name_variants.yaml`.
+- Bugfix: In first names, the values `None` and `""` (empty string) are now considered equivalent.
+- Bugfix: Name variants in different scripts are now correctly recorded as names for the respective author.
+- Bugfix: `MarkupText.as_html()` now always correctly HTML-escapes characters.
+- Bugfix: `MarkupText.from_xml()` now correctly handles empty tags; got converted to the string `"None"` before.
+
 ## [0.4.3] — 2023-11-05
 
 ### Added